requirements.txt
apache-airflow-providers-cncf-kubernetes
google-api-python-client
pyspark
boto3
